[类别管理] 0824  -09:24

+++ spring-cloud-woniu-manage/brand-service/src/main/java/com/wcq/brandservice/dao/mysql/BrandDao.java
@@ -12,5 +12,5 @@ public interface BrandDao extends BaseMapper<BrandPo> {
 
 
 
-    List<BrandDto> getBrandInfoByid(Integer id);
+    List<BrandDto> getBrandInfoByid(Long id);
 }
